make agent detail sections
