+++ web-app/src/components/agents/details/index.ts
@@ -1,0 +1,5 @@
+export * from "./section-1";
+export * from "./section-2";
+export * from "./section-3";
+export * from "./section-4";
+export * from "./section-5";
